fix(plugins): break update type cycle
