+++ src/plugins/update-source.ts
@@ -1,5 +1,7 @@
 import { isRecord } from "@openclaw/normalization-core/record-coerce";
+import type { OpenClawConfig } from "../config/types.openclaw.js";
 import type { PluginInstallRecord } from "../config/types.plugins.js";
+import type { ClawHubTrustErrorCode } from "../infra/clawhub-install-trust.js";
 import { parseClawHubPluginSpec } from "../infra/clawhub-spec.js";
 import { satisfiesPluginApiRange } from "../infra/clawhub.js";
 import { unscopedPackageName } from "../infra/install-safe-path.js";
@@ -36,7 +38,74 @@ import {
   resolveOfficialExternalPluginInstall,
 } from "./official-external-plugin-catalog.js";
 import { resolvePackagePluginApiRange } from "./package-compat.js";
-import type { PluginUpdateChannelFallback } from "./update.js";
+
+/** Logger surface used by plugin update flows. */
+export type PluginUpdateLogger = {
+  info?: (message: string) => void;
+  warn?: (message: string) => void;
+  error?: (message: string) => void;
+  terminalLinks?: boolean;
+};
+
+type PluginUpdateStatus = "updated" | "unchanged" | "skipped" | "error";
+
+export type PluginUpdateChannelFallback = {
+  requestedSpec: string;
+  usedSpec: string;
+  requestedLabel: string;
+  usedLabel: string;
+  reason: "unavailable" | "failed";
+  message: string;
+};
+
+type BasePluginUpdateOutcome = {
+  pluginId: string;
+  message: string;
+  currentVersion?: string;
+  nextVersion?: string;
+  channelFallback?: PluginUpdateChannelFallback;
+  warning?: string;
+};
+
+export type PluginUpdateOutcome =
+  | (BasePluginUpdateOutcome & {
+      status: "skipped";
+      code?: ClawHubTrustErrorCode;
+    })
+  | (BasePluginUpdateOutcome & {
+      status: Exclude<PluginUpdateStatus, "skipped">;
+      code?: string;
+    });
+
+export type PluginUpdateSummary = {
+  config: OpenClawConfig;
+  changed: boolean;
+  outcomes: PluginUpdateOutcome[];
+};
+
+export type PluginUpdateIntegrityDriftParams = {
+  pluginId: string;
+  spec: string;
+  expectedIntegrity: string;
+  actualIntegrity: string;
+  resolvedSpec?: string;
+  resolvedVersion?: string;
+  dryRun: boolean;
+};
+
+export type PluginChannelSyncSummary = {
+  switchedToBundled: string[];
+  switchedToClawHub: string[];
+  switchedToNpm: string[];
+  warnings: string[];
+  errors: string[];
+};
+
+export type PluginChannelSyncResult = {
+  config: OpenClawConfig;
+  changed: boolean;
+  summary: PluginChannelSyncSummary;
+};
 
 export type UpdatablePluginInstallRecord = PluginInstallRecord & {
   source: "npm" | "marketplace" | "clawhub" | "git";
